split App/AppContent for lazy loading
